Minor proofreading edits
Love the piece! Used to edit tech books for a living and so proofreading as I go is second-nature, and since it's on GitHub, figured I might as well make the edits as I read them. Mostly just fixed some typos and cleaned up the style to be more consistent throughout the piece (for example, in terms of when to use the code font vs. when to use normal text, or whether punctuation is included in links). 

There are two (still very small) content changes I made that are worth explaining. First, in the example about dependencies, I changed the NumPy reference to NLTK because the latter is what you showed in the code example immediately below it (it was confusing otherwise). Then, in the unit testing section, I swapped the reference to nose out for nose2 because nose's own documentation page says you should use nose2 or unittest instead because it's not going to be maintained (so probably doesn't make sense to tell people who don't already use it to start).